backend/requirements.txt
fastapi
uvicorn
ccxt
python-dotenv
python-telegram-bot
motor
tensorflow
pandas
scikit-learn
apscheduler